Add shorthand access to unit sentries
